+++ src/Products/PloneMeeting/tests/testUtils.py
@@ -18,6 +18,7 @@ from Products.PloneMeeting.ftw_labels.utils import get_labels
 from Products.PloneMeeting.tests.PloneMeetingTestCase import PloneMeetingTestCase
 from Products.PloneMeeting.utils import duplicate_portal_type
 from Products.PloneMeeting.utils import escape
+from Products.PloneMeeting.utils import isPowerObserverForCfg
 from Products.PloneMeeting.utils import org_id_to_uid
 from Products.PloneMeeting.utils import sendMailIfRelevant
 from Products.PloneMeeting.utils import set_dx_value
@@ -361,6 +362,40 @@ class testUtils(PloneMeetingTestCase):
         self.assertEqual(get_labels(item), {'label': 'Label', 'suivi': 'Suivi'})
         self.assertEqual(get_labels(item, False), {'label': 'Label'})
         self.assertEqual(get_labels(item, "only"), {'suivi': 'Suivi'})
+
+    def test_pm_IsPowerObserverForCfg(self):
+        """ """
+        cfg = self.meetingConfig
+        self.changeUser('pmManager')
+        self.assertFalse(isPowerObserverForCfg(cfg))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['powerobservers']))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['restrictedpowerobservers']))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['powerobservers', 'restrictedpowerobservers']))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['unknown']))
+        self.changeUser('powerobserver1')
+        self.assertTrue(isPowerObserverForCfg(cfg))
+        self.assertTrue(isPowerObserverForCfg(
+            cfg, power_observer_types=['powerobservers']))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['restrictedpowerobservers']))
+        self.assertTrue(isPowerObserverForCfg(
+            cfg, power_observer_types=['powerobservers', 'restrictedpowerobservers']))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['unknown']))
+        self.changeUser('restrictedpowerobserver1')
+        self.assertTrue(isPowerObserverForCfg(cfg))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['powerobservers']))
+        self.assertTrue(isPowerObserverForCfg(
+            cfg, power_observer_types=['restrictedpowerobservers']))
+        self.assertTrue(isPowerObserverForCfg(
+            cfg, power_observer_types=['powerobservers', 'restrictedpowerobservers']))
+        self.assertFalse(isPowerObserverForCfg(
+            cfg, power_observer_types=['unknown']))
 
 
 def test_suite():